move downing strategies to separate file
