+++ akka-cluster/src/main/scala/akka/cluster/sbr/SplitBrainResolverSettings.scala
@@ -1,6 +1,7 @@
 /**
  * Copyright (C) 2015-2020 Lightbend Inc.  <https://www.lightbend.com>
  */
+
 package akka.cluster.sbr
 
 import java.util.Locale
